Y axis accessor function adjustment
The quality value data is stored as a string in the json object, and so we need to parse the quality value as a float, otherwise the strings are compared, and this can lead to problems when d3-array extent function tries to calculate minimum/maximum. For example, it will consider "9.83" to be larger than "10.0" which causes the Y axis to fly all over the place.